chore: extract docker host calculation to an internal package
